change dir

+++ site_tools/utils/catalog/catalog_upload_v3.py
@@ -8,13 +8,12 @@ import zipfile
 import xml.etree.ElementTree as ET
 from site_tools.utils.catalog.object_create import ObjectCreate
 from site_tools.utils.catalog.object_update import ObjectUpdate
-from site_tools.utils.catalog.report_create import create_report
-from site_tools.utils.catalog.constants import report_columns_dict, report_message, stop_list_product
-from site_tools.utils.error_logs.create_db_massege import create_message_db
+from site_tools.utils.report.report_create import create_report
+from site_tools.utils.constants import report_columns_dict, report_message, stop_list_product
+from site_tools.utils.messages.create_db_massege import create_message_db
 from django.core.exceptions import ObjectDoesNotExist
-from django.db.utils import OperationalError
-from site_tools.management.commands.dump_catalog import dump_data_catalog
 from django.core.management import call_command
+
 
 class TempCategoryUpdate(ObjectUpdate):
     model_django = Category
@@ -261,35 +260,42 @@ def transaction_upload_catalog():
     """
     catalog = None
     report = None
+    error_message = None
 
     with transaction.atomic():
-
         try:
             catalog = Catalog.objects.get(pk=1)
         except ObjectDoesNotExist as e:
-            create_message_db(f'Каталог не обнаружен -> {str(e)}')
+            error_message = f'Каталог не обнаружен -> {str(e)}'
+            create_message_db(error_message)
 
         Category.objects.select_for_update().all()
         Group.objects.select_for_update().all()
         Product.objects.select_for_update().all()
         call_command('catalog_dump')
-
         try:
             report = CatalogUploader(catalog).temp_catalog_parser()
-        except OperationalError as e:
-            create_message_db(f'Ошибка при обработки каталога -> {str(e)}')
-            catalog.status = f'Ошибка при обработки каталога -> {str(e)}'
-            catalog.save()
-
-    catalog.status = 'Готовлю отчет'
-    catalog.save()
+        except ValueError as e:
+            error_message = f'Каталог не загружен. Каталог -> {str(e)}'
+            create_message_db(error_message)
 
     try:
+        catalog.status = 'Готовлю отчет'
+        catalog.save()
         create_report(report)
     except Exception as e:
-        create_message_db(f'Ошибка при попытке создать отчет -> {str(e)}')
-    catalog.delete()
-    create_message_db("Каталог успешно обновлен и файл удален, подготовлен отчет и резервная копия.", notification=True)
+        create_message_db(f'Отчет по каталогу не создан. Отчет -> {str(e)}')
+
+    if error_message:
+        create_message_db(error_message)
+        catalog.status = error_message
+        catalog.save()
+    else:
+        catalog.delete()
+        create_message_db(
+            "Каталог успешно обновлен и файл удален, подготовлен отчет и резервная копия.",
+            notification=True
+        )
 
 
 def transaction_upload_catalog_test(cat):
